DEPR: __invert__ on object dtype (GH#51567)
~ on object dtype dispatches elementwise to the objects' own __invert__,
so a boolean mask that has been cast to object silently becomes integers
instead of being negated.

Object is the only dtype left to deprecate: ~ currently succeeds only on
bool-like, integer-like, and object dtypes. datetime64, timedelta64,
period, interval, category, float, complex, string, decimal and every
non-integer arrow type already raise.

+++ pandas/tests/indexes/ranges/test_range.py
@@ -867,7 +867,9 @@ def test_invert_returns_rangeindex(rng):
     ri = RangeIndex(rng, name="foo")
     result = ~ri
     assert isinstance(result, RangeIndex)
-    expected = ~Index(list(rng), name="foo")
+    # pass dtype explicitly: an empty list would otherwise infer object dtype,
+    #  for which ~ is deprecated (GH#51567)
+    expected = ~Index(list(rng), dtype=ri.dtype, name="foo")
     tm.assert_index_equal(result, expected, exact=False)
